PC uart communication tasks and circular buffer library is written.
PC_Communication related tasks are completed.
IWDG is disabled for now.
A simple circular buffer library is written and used for UART data coming from PC (usart2).
System is working and tested with Python (test script will be put here in GitHub soon)
A readme is added to firmware folder.